fix typos
The words 'canceled' and 'cancellation' can both be spelled with either one or two 'l's (depending on whether you use American or British English). In System.Threading.CancellationToken 'canceled' is spelled with one 'l' while 'cancellation' is spelled with two 'l'. The spelling here is just to be consistent with that.